+++ sdk/monitor/azure-monitor-opentelemetry-autoconfigure/src/test/java/com/azure/monitor/opentelemetry/autoconfigure/implementation/pipeline/TelemetryItemExporterTest.java
@@ -68,7 +68,7 @@ public class TelemetryItemExporterTest {
         TelemetryPipeline telemetryPipeline = new TelemetryPipeline(pipelineBuilder.build(), null);
 
         return new TelemetryItemExporter(telemetryPipeline,
-            new LocalStorageTelemetryPipelineListener(50, tempFolder, telemetryPipeline, null, false));
+            new LocalStorageTelemetryPipelineListener(50, tempFolder, telemetryPipeline, null, false, null));
     }
 
     private static String getRequestBodyString(Flux<ByteBuffer> requestBody) {